Intelicarding an AI no longer causes two sets of AI laws to be reported at round-end.
You can now view the laws of an Intelicarded AI, and initiate a wipe of its files, killing it. Further functionality, such as giving it the ability to talk over a radio, is planned.

It is no longer possible to use an Intelicard on an AI core that has already had its AI removed.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@124 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 6c581cced457621094060931f57f1b75481b26ae