Bridge changes.
Eliminated the command centre in the middle of the bridge, it was bulky and despite being useful didn't really do much in the way of good for the bridge. 

Instead added three departmentalized stations to the bridge to better organize the bridge. Also added two status terminals to the front. Stations are:
Security station - Security records and cameras. 
Conn / Operations - Communications, cameras, and station alerts. 
Medical - Crew monitoring and medical records.

I ran this by as many people as possible to get feedback, but worst case scenario we can always revert to the old bridge.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@719 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 726416d66fea99af438dc9d2aa5f136f656d24a6