AIRLOCK update: - fixed shock code (code by barhander) - made them spark even if you're wearing insulated gloves. does no damage to you tho. - fixed them sparking if no power is available
MAP updates:
- Slightly expanded mining dock's construction site. added an area for it and an APC. named it 
- Completed the turfwork for the engineering construction site, gave it an area, an APC and connected it to the scrubber and vent pipeworks.

I've been moslty doing 15 minute projects the past week due to schoolwork. this one wasn't such a project, but i felt i owed to put more teme into it, as it was my map update which exposed the coding error for airlocks. (altho the error was not really my fault, it was still my update which exposed it.)

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@916 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: f6d380fc76bd3050e24c2d41f6aaaa9a03d048dc