MMI update the third and fourth. Barring some big huge bug that didn't turn up in my testing to make sure shit works, the MMI is ready for showtime.
If you're wondering why updates third and fourth happened together, in update the third I redid how brains work, and I was halfway to committing it when I realized "wait, if I upload this now, everything that uses the brain will cause fuckups".

Notable changes:
When your brain is cut out, instead of your client hanging around inside your now brainless and more or less useless body, it it shunted into the brain object, similar to an AI in an intellicard.
While not in an MMI, your brain is "dead", deathchat and all. When stuck into an MMI, you spring to 'life', allowing you to talk to whoever stuck your brain into the thing. It occurs to me while I'm typing this up that it would be very easy to abuse this communication between life and death by slapping brains in and out of MMIs. DO NOT DO THIS OR THE ADMINS WILL HATE YOU AND DO MEAN THINGS TO YOU.

Life in an MMI is pretty boring, so you should yell at people until they stick you into an AI core or robot frame to give you something to do. I know the MMI was originally thought up as a mecha interface, but I haven't gotten around to looking at the mecha code to figure out how to hook it in yet.

I suspect there's going to be howls of "WHY DO YOU HATE ROBOTICISTS" over this next part, but the MMI itself is an R&D-researched object, with a biotech level of 3. That said, the station starts with two of them in the operating room (to get the research level up if nothing else), and they are producible at the protolathe AND the mecha fabricator (since they will eventually get to plug into a mecha.)

Known issues: The MMI is SUPPOSED to pop out of a cyborg with the player attached when the cyborg goes pop, but it didn't in my initial testing, simply ghosting the borg as they already do. Will work on this as I have time, but what with having to go to work these days, I have somewhat less of that.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1311 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 9e50fd8e962e0b9d4d32fb08eaa1565215e45902