BORG VISION MODE UPDATE
Added three new items for use in borg modules.

Meson Vision
Thermal Vision
X-Ray Vision

When selected as an active module (as in it is taking up one of your three module slots, not that it is your "in-hand" module), the module confers the appropriate vision mode. These vision modes consume a lot of power while active (50/tick on meson, 100/tick on thermal, 200/tick on xray), so it's best to keep them inactive unless you need them.

Gave Meson vision to mining and engineering borgs, did NOT hand out Thermal or Xray vision, that won't happen until I get an idea of public opinion on this update.

Also included is a syndicate borg module that's been sitting half-finished in my local copy of the code as I was having issues deciding what all to give it. It is NOT selectable at this time. The plan was for it to only be selectable under specific conditions,  most likely involving hacking and/or emagging. It's only getting committed because its in the same file as the actual update.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1359 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 6360a1764938058e38c53b638b397813ce93e5c8