Chucking my powered armor onto the SVN
It could use a fair amount of refinement and/or additional features yet, but it is functional. It is currently adminbus only, until I get some non-placeholder sprites. Already have some on the way from a pretty decent spriter, but don't let that discourage you, more types of power armor are always welcome.

Current model is fueled by processed plasma ore from the mine. Each unit of plasma provides roughly 2 minutes of power, the tank holds 2 units worth of fuel at a time, and can be refilled while the suit is active. I'll probably include a powercell and self-sustaining version when I add in the code to make this sucker buildable via R&D

While unpowered, the suit slows you waaaaaay down, is not spaceworthy, and is inferior to regular armor. While powered, it is between regular and syndicate spacesuits in speed, can be worn in space, and has much better armor. Also, while powered it cannot be unequipped, much like an initialized ninja suit. The helmet is required for the suit to be able to power up, and if the suit powers down for any reason other than the user initiating the powerdown sequence, the suit does NOT unlock, so you're stuck in a crappy slow metal suit until you can get it powered up to properly power it down.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1658 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 02c951575dd2932057ff99bc5ddff53979e1301e