I added in three new clothing pieces. A chicken suit, a Holidaypriest suit and a nurse suit and nurse hat.
I added the holidaypriest suit and the nurse suit and hat to their respective lockers in chapel and Medbay. I also placed the chicken suit in the theatre.

Barman has a new drink named Singulo and Carrot Fries and Meatball spagetti was fixed.

Thanks goes to Farart for all these great sprites. So thanks Farart.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1806 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 988f795be2dee02a8252d3e584fbb94d43d4a965