Added 5 new plants to hydro. Two of them are tomato mutations. A blood mushroom and the blue mushroom. You can get blood from the blood mushroom, and it will soon be used in some new recipes. The blue mushroom is a surprise you will have to see.
I added two new types of berry mutations. One is death berry, which is actually a mutation of the poison berry. It is EXTREMELY deadly if ingested. Another is the glow berry. Its brightness depends on its potency.

Last, but not least. We have a plant which is both a hydro and harvestable plant, but at the same time a livestock. It is the Walking Mushroom. It is a mutation of the plump helmet. After it is harvested and in your hand, click it to bring from the Walking Mushroom. If it dies, then use a knife on it for Huge Mushroom Slices.

Next commit will have more recipes for both chef and barman.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1843 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 625c503db2c2f077c91f1455a0b3872622e16b89