Damage system changes: - Environmental heat damage increased by a factor of 1.5 (meaning space and fires cause 50% more damage) - Being without oxygen causes from 0 to 12 damage instead of 0 to 7, depending on how much oxygen is in the atmosphere. Means that if the room's just began to leak there will be no change in the oxygen damage you take compared to the damage prior to this update, but space does 12 damage per tick instead of 7. - Fire suit now protects up to a temperature of 10000 degrees, which is the temperature at which floors start to melt. A toxin + oxy fire usually generates temperatures around 8000-9000 degrees. It used to protect up to 4500 degrees. - Added a standard spacesuit loadout to the admin select equipment menu (LJ's request)
I've tested all the changed numbers a lot and they seem to work well. Space and fires are a lot more deadly, while normal circumstances remain at the same level as they are now.

I've also researched the bug which allowed you to spend infinity on tiles in space. The problem is that turfs don't process temperature properly, but even after spending 4 hours looking into it (I wish I was joking), I wasn't able to determine where this gets updated on the lowest of levels.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1896 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 044ab25cf8a1d4c28e0fd457e4de87eb446f501b