Users could dupe diamonds. I fixed that.
Insert rate was 3750 and eject rate was 1000. 

You could eject a sheet from a protolathe and get 3750 cm worth, but only 1000 cm worth was taken out of the machine.
Repeat this and you can turn 1 sheet into 50 in no time. So BUGFIXED IT. 

Credit to Midaychi for finding this bug.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1959 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 937543c32eb89ead558fb262e22d3de6e8a22aec