Finally got around to working on that power armor overhaul I'd told myself I was gonna do.
tl;dr is that instead of being a single object, power armor now has component objects, which determine how good it is at things. They are still a work in progress, lacking a way to actually construct them and such. That said, there are adminspawnable sets. One is more or less what the old power armor was, sans a runtime error that was causing it to be immune to everything, the other is an adminbus set that is genuinely immune to everything.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2023 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: d867f56b75fe314069884dca24837e9ed29fe58b