We decided to give R'n'D a table facing the hallway. Hopefully this will mean ordering stuff from r'n'd will be easier.
To do the above, I swapped the location of gas storage and R'n'D. I admit R'n'D is not the prettiest of rooms but eeh, whatever.

Now onto the stuff that'll get this rev downvoted, dispite me saying I want to keep them, people (the ones who have influence) wanted me to remove the pair of insulated gloves in research maintenance. I also removed the two fire suits, mesons and 4 of the 5 utility belts in the same hallway. Multitool stays.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2098 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 96d74134410ba7309d3d0e010ac2d73be092c251