Updated the interface a bit to allow a built-in browser.
The upper right panel, where the verbs usually are now has two functions, verbs and the browser. This feature has been added to hopefully reduce the number or pop-ups that appear. Currently only request consoles and vending machines use this feature. I'll wait for feedback and then update the rest, if feedback is positive.

Also removed the redcoat spawn pack... The redcoat suit is MINE.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2170 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: e2ffd2f5b2537c99ba8956204af4425b032cc2a4