- "delete all of type" added to the view variables panel.
It requires you to first select whether you want to delete all objects of the strict type or of the type and all subtypes. After this you need to confirm. And then you need to confirm again. If anyone manages to get through 3 confirmation screens 'by accident' they are lying.

You can only use 'delete all of type' for objects (of type /obj)

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2265 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 3794c0c64fd7437c9dfbf9c7f9bed247b99f5104