Escape pods added to test out the concept.
Before you read on this revision is more of a thing being tested than a feature that is guaranteed to stay.

4 escape pods have been added to the station, 2 at arrivals, 1 in engineering and one in security. Each pod is 4x3 tiles in size with a window tile, 2 chairs and a door. So it has 3 floor tiles to stand on.

Pods don't have any sort of emagging capabilities, hijacking a pod does not count as completing the escape alone objective, surviving in a pod however counts towards the escape alive objective. Cult members in pods do not count towards the acolyte victory.

Changelog updated with listed changes.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2663 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: e7bea1cfb04e35b11507d7f89a2fab4ca69d3ffd