Hopefully made meteor mode more deadly.
Fixed meteor spawning so all the meteors that are set to spawn actually hit, but noticed that 50 meteors every 5 seconds proved itself to be way too laggy when all of them actually spawned properly and aren't deleted at soon as they're created due to being on the map's edge. Instead I lowered it to 6 per second.

Wrote this in the changelog.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2679 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: b04400e11c6a24f0ef670c3dcdf9aa63ad1b12fa