New verb: rightandwrong
When used, it will arm every human mob that is not already a special role (and with a client assigned) with a random gun from the following list:
-Taser
-Egun
-Laser
-Revolver
-C20r
-Decloner
-Desert Eagle
-Gyrojet Pistol
-Pulse Rifle
-Silenced Pistol
-Laser Cannon
-Combat Shotgun
-Mini-Uzi
-Mini-Ebow

It will then inform the player that
"You are the survivor! Your own safety matters above all else, trust no one and kill anyone who gets in your way. However, armed as you are, now would be the perfect time to settle that score or grab that pair of yellow gloves you've been eyeing..."

And make them a traitor/give them the objective to survive till the round ends.

After use, rightandwrong will remove itself from the users list of verbs.

rightandwrong addeded to the Spellbook under "Summon Guns"
The verb will appear in the spell tab.

Updated changelog

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3071 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 7321826910f1fa1f116c20622bde11ee63abc909