Fixed projectiles going into disposals.
Made the rapid syringe gun a subtype of the syringe gun, and reduced it's capacity to four syringes.
Fixed the examine bug on syringe guns.
Added a new sprite for the regular syringe gun.

Added a new door for research. It's also used in robotics.

Updated the air injector sprite.

Fixed some pepperspray spelling errors.

~~REMOVED~~: the energy crossbow and chemsprayer from R&D.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3138 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: f41b5cbabb5ca858bc51119e2dc507a66eb59cfa