Holy crap BYOND is retarded. ;-;
\icons[object.icon] does not work correctly and may cause client crashes. This totally doesn't happen when testing locally so I guess the icon stuff must be done client-side/be lag-induced or something? I've also noticed that icon-updating is very slow (items tend to disappear briefly when being made bloody looking). With the recent forensics updates everything gets bloody, resulting in massive bulshit.

I am working on a rewrite of bloodied-icons. For now, this should stop runtimes from examining humans caked in blood. Unfortunately the little icons in the chat-log will no longer be visible when items are bloody, I hope to fix this with said rewrite of bloodied-icons.

Moved the paper_talisman icon into bureaucracy.dmi to make Cult mode playable.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3277 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: e791749512c7e8c138b14a7eb43e93544cfeb2e9