TG: Added two new variables, "canstun" and "canweaken". They can be modified in runtime to prevent mobs from ever being weakened and stunned by tasers/batons/electrified doors.
Xenos are immune to stunning and weakening. This basically means you cannot run
up to a xeno and shoot them with a taser and game over. This makes them a lot
more dangerous. To compensate, I've added in a xeno-exclusive variable
controlling movement delay addition. When a xeno is stunned, its movement speed
goes down. The speed is recovered in the life() proc.

Modifications to critter.dmi (manhacks). Added a new lighting icon, floor
lights. Possible implementation later. r2886


Former-commit-id: dd2b3cc816b6ec6fbe51154ff9adbf1b320b7b45