TG: (Does not actually remove c20r) Desert Eagles now use magazines, .50AE ammo, and only have 7 shots. They're basically revolvers that use magazines.
Nuke ops lost that horribly nerfed piece of shit c20r and now get Desert Eagles
and Riot Shields.

The detectives .38 ammo now does the same damage as regular revolver bullets. If
he wants to keep acting as a glorified sec officer, he can go grab a taser.

The Veil Render now summons a subtype of Nar-Sie which doesn't pull anchored
objects. This should reduce the power of the item/reduce lag while I get sprites
and finalize details with Urist for reworking it as the Sphere of Annihilation
(as in, this is temporary)

Updated the changelog


Former-commit-id: e174e7ca76d196fd0c5a7455ce5de5016d85d3a5