TG: Fixed projectiles going into disposals.
Made the rapid syringe gun a subtype of the syringe gun, and reduced it's
capacity to four syringes.
Fixed the examine bug on syringe guns.
Added a new sprite for the regular syringe gun.

Added a new door for research. It's also used in robotics.

Updated the air injector sprite.

Fixed some pepperspray spelling errors.

~~REMOVED~~: the energy crossbow and chemsprayer from R&D.


Former-commit-id: 60caca6a2a1ea88ea96c7d4ed29ec9935701f707