TG: Disposal and welding helmet exploit fixes:
- You can no longer toggle the welding helmet when stunned or restrained.

- You can no longer give yourself all-access via-disposals. Living mobs and
living mobs inside of a container cancel their holder's destination. This still
allows people wrapped in lockers to be sent to their proper destination, however
it would require someone on the other end to open it. So it's their fault for
not dragging it out before opening it.
Revision: r3400
Author: 	 johnsonmt88

Former-commit-id: d6d3d60e0dce3e17c86a26990a3c8635069f93af