Changed the way facial scarring works. Instead of setting your real_name to "Unknown" it uses the disfigured variable of a human's head organ to dictate the human's name variable.
This means real_name is now somewhat back to being a reliable source of the mob's actual name. It should eliminate a lot of the "cloning as unknown" bugs.

It also means I could simplify that god-awful name updating stuff into a nice and simple helper proc.

Some original_name stuff was added here and there,  mainly during cloning. A lot of the "getting random ghost names" should  be fixed now. Still loads to do though, particularly with transforms and such. >_>

Fixed a runtime with Tajarans trying to use a variable that doesn't exist for PDAs. Removed that variable from IDs as it's only used by furries.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3546 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: d835287ddf2d28e5499de28c465fe70324cd8ccc