TG: Rewrote the cinematic system to be centralised and controlled by the gameticker, rather than handled per mob with like, 3 loops and a seperate process for each mob.dmPlease report any bugs/derps to me asap. I've tested it as much as I can locally, but because it relates to events that require a large number of players it's very hard to test.
TODO: I was intending to use dust to kill players off on a nuclear-loss. However
dust() is very very broken. So I've hardcoded everything to die until I can fix
dust().

Added an adminverb for 'Game Master's. It was a debug verb so I could test
cinematics but I've left it in as it might be handy for events, especially if we
get more cinematics. To access it type "cinematic" in full into the input bar as
a Game Master, then follow the prompts. At the moment, it only handles nuclear
stuff (that's all there is anyway). You can choose where the nuke detonated 0=
on station; 1= near station but in space; 2= off the z-level. You can also force
a cinematic from another game-mode to play with the prompt after that one.

Note: All mobs are buckled to a bed located inside the gameticker (odd I know)
whilst a cinematic is playing. The bed is deleted afterwards releasing all
players. If off_station=0 then all mobs will be killed too. This is to stop
people running around and doing stuff during the cinematic. I really didn't want
to add more variables/checks to mobs.
Revision: r3481
Author: 	 elly1...@rocketmail.com

Former-commit-id: bec822a04930184cf88ec5ced433fc4131ea3bd2