Fixed issue 411. (Thanks Yvarov- http://nanotrasen.com/phpBB3/viewtopic.php?f=16&t=8651) Added Invisty's new field generator sprites.
Added a button to check antagonists that stops the world from automatically restarting at round end.
It's toggleable, but once the round ends with it on, you'll need to reboot the world manually. (It's under check antagonists because that's where the shuttle stuff is, let me know if I should move it.)

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3565 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: e89e7acdb3ca56ff0ab85ab29dcbbbb18d7824dd