Updated toilets, urinals, and showers with NEW features.
The shower path has changed!
From /obj/structure/shower to /obj/machinery/shower

Made bookcases appear empty if they have no books, and fill up as more books are placed in them.

Updated do_after() to have an extra var that determines if the user is required to keep the item in the active hand. This is only used in swirlies at the moment.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3684 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: f2f3c6bebf7d71efc0a0fdf887d7977ebddeee86