Fixes a derp I made with the PA in a recent commit.
Rules verb moved to the interface to make it more visible (so new players may actually -read- it occasionally).

Minor changes to investigate singulo based on feedback. Entries are now timestamped and the "time \ref[src] (location)" bit is in a smaller font. Collectors note how much fuel they are loaded with. 

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3687 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 7f526052773ffa1bb1d70960ecaa19569b92317d