Fixed r500 (faces displaying above glasses and such) Cause: face layer was set to MOB_LAYER. Fixed r501 (underwear not showing) Cause: bad indentation
A possible fix for some very odd bugs caused by restarting the MC related to the usr variable being passed into the MC. *shrug* Very hard to test.

Removed the need for the world to be in Debug2 mode to edit ticklag as the verb is hidden behind another proc now anyway, and it's become more of an annoyance than a thing to stop people accidentally clicking it.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3695 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 40c2e25bde9fd5446f7c9d223cea36ea87f646db