Changed a bunch of constants to defines and made setup.dm all pretty and easy to read.
Defined some clothing-overlay layers. The defines are float-layers. So they will always appear above the mob they are overlaying. They can be found in setup.dm

This pretty much means you can change to order in which the icons are added to the overlays variable without fear of eyes appearing above glasses and such. So no more of these dumb bugs.
It also paves the way for selective clothing updates which should speed things up a fair bit.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3704 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 1b1ddac33b19a68bc31d1a1a9d6432a8a04ccc65