TG: Changed a bunch of constants to defines and made setup.dm all pretty and easy to read.
Defined some clothing-overlay layers. The defines are float-layers. So they will
always appear above the mob they are overlaying. They can be found in setup.dm

This pretty much means you can change to order in which the icons are added to
the overlays variable without fear of eyes appearing above glasses and such. So
no more of these dumb bugs.
It also paves the way for selective clothing updates which should speed things
up a fair bit.
Revision: r3704
Author: 	 elly1...@rocketmail.com

Former-commit-id: 65416c202f438aecb4619934d378bdef9b24f743