TG: Comitting for Giacom. Details here: http://nanotrasen.com/phpBB3/viewtopic.php?f=16&t=8960
And, to summarise:
Added a new computer, the message monitor console. It is able to access the logs
on the message server.
Made PDAs dependant on the message server instead of Telecomms. This fixes PDAs
not sending any messages due to no listening_freqs.

When paper is splashed with Ethanol, the contents of the paper or book
disappear.
Added white coloured pens. Allows you to write hidden messages that can be
revealed by highlighting the text.
Revision: r3777
Author: 	 petethegoat

Former-commit-id: 4f4ebd809d8f7c1d176b2d87230ae9785d9b3b79