Added some restrictions to round-start names for humans. this was necessary due to people using meta names like "floor" "unknown" "lol" etc.
If you feel the restrictions are to strict please let me know here or on the feedback forums.

Generally only A-Z, a-z, space, . - ' are allowed.
Names should be 2 words at least.
Names with too many space, . - ' are rejected.
Names are auto-capitalised.
Names with not enough regular alpha-numeric characters are rejected.
Names with bad characters such as \ etc are rejected.

Raised the minimum character age at the request of a few people. It's now the minimum age for enlisting in the army for most countries (17)

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3956 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 2a775f9d1f9d2e1199fe61eb2ddbfdf8c53427db