Committing this in-case anybody is willing to speed-test it against the current one.
It's an explosion proc that uses locate to locate only turfs and then ex_acts them. As opposed to grabing every atom in a square around the epicenter and then selecting which ones to ex_act

this version also supports sticking a sleep() in there to help  stop the server getting too backlogged.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4134 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: f40dcf358e8d351219568b05293b3bff5e21394f