Mimes can no longer whisper without breaking their vows. Mimes don't talk!
Lockers and crates now holy a maximum of 20 objects. 
This is to prevent players from dragging a locker into a populated area and opening it; resulting in clients with a weaker connection to lock up and crash. I realize this is an inconvenience for botanists but players being able to intentionally crash other players takes priority.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4195 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 7d1b47e4820814753d69eeefadccc1a125a81f7b