Tidied up on_consume() and the trash code for snacks.
Trash is now the path for the trash-item prototype (the thing it will spawn when you finish eating). This removes the need for a switch. Only obj/item/ paths may be trash.

You can also set var/trash to any obj/item to dump that item as trash. This is now used by the fortune-cookie rather than giving it its own proc for no reason.

This resolves issue 687

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4201 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 92dd10f89283819494c41b1b728e5ca6d2f17a79