It is now easier to modify UI's through genetics:
Instead of the subblock picking a random value far from the current value, the player can select a 'target' value to aim for and subblock will try to zero-in on that target.

This affects UI's only, SE's don't do this. This will not be added to SE's because superpowers don't need to be made easier to obtain.

Hitting the arrows to change blocks/subblocks will now loop around once you've hit a limit. 
- For example if you're on SE block 14, clicking higher will bring you back around to block 1.

Commented out the scan console process() as it was just doing meaningless checks.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4204 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 24ec69ac7ae6f6595620361386ab4646af873630