Added the warden jacket and pepperspray to the warden's wardrobe, and added fingerless gloves to the cargo technician and shaft miner wardrobes.
The janitor spawns with a portalathe but doesn't have one in his wardrobe, so this adds that. If the librarian gets a barcode scanner in his, then the janitor's should include the portalathe.

The captain has taken a few lessons from the head of security! When joining, the captain will now start with matching gloves, a pair of jackboots, and the more sensible captain's cap. The old Napoleonic hat is still available in the captain's closet, and both the closets and wardrobes were updated to reflect this change.

The quartermaster started with brown shoes, but only had black ones available in his wardrobes. This change should make his starting equipment more consistent.

The head of personnel should no longer be so paranoid, and now spawns without wearing body armor or a helmet. Additionally, the body armor and helmet from the wardrobe closet and bag were removed. A gun, helmet, and body armor are still available in the secure locker for emergencies.

The breath mask inside the engineer box included with the ERT backpack looked terrible when worn with the ERT helmet, yet was required if any of the team were operating near a breach or anything. This change only returns the removed SWAT mask (which leaves the helmet sprite intact) to the response team's basic equipment.

Both the secure closet and the wardrobe bag have the medsci headset in it, and the chemist lab is conjoined with research, so why shouldn't the chemist start with that headset? This commit also moves the chemist to Medical on the roster because that's where the job shows up everywhere else.


Former-commit-id: eac13356079dced2ba1ac3f679eec262080eebf1