Fix for heads not getting assigned properly at round start.
Well, hopefully a fix. I've tested it locally to make sure it still actually assigns jobs without any problems, and it should work but we wont know for sure until we get more people trying it out.

Thanks to Tonberrytoby for correcting me and pretty much pinpointing the source of the problem.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4296 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 8ae701bd014da42f4e954869add74edb8a7ea0f8