Fixed issue 777. Fixed issue 439.
Relating to the 439 fix, added a new define, TEMPERATURE_DAMAGE_COEFFICIENT. The name is slightly misleading, as it is used in reagents that affect body temperature. Leporazine now functions properly, and you can once again into space with a cup of coffee and a firesuit.

Also did a few grammar changes for newscasters.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4392 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: e4e7256d66fffa43aba122fe4facdbd258266753