Split the gateway/center into two objects, stationcenter and stationaway. The gates actually search for and detect eachother, and warp to eachother (suck it Pete)
The away gate doesn't require power to make away mission mapping possible in unpowered areas like caves, the station gate requires less power than it did (for the time being, it'd be nice if people could actually use the gate. Makes no sense to make new content so hard to use. If it becomes a problem, I can up the power requirement again).

Added a "calibrated" var to the away gate. If it's uncalibrated, you'll warp to the destination landmarks. Hitting the gate with a multitool with calibrate it, and any more warping to the gate will be on target.

It defaults to calibrated, so remember to set it to 0 if you want people to appear somehwere other than at the gate.

Updated Cheridans map with the new gate type (and made it uncalibrated)

Updated the example map to use the new gate type

Updated the station to use the new gate type

Changed the gateway description to reflect it's unknown origin

To do: Make gates able to block sending from the other gate (so you can exile people), make damaging gates so they can't send/recieve (and still do the other) possible. Give the station gateway unique sprites (for Nanotrasens own prototype gate, with appropriate description)

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4718 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 84308d40426bf09f3b649272eed8debb7e8e55d3