Reduced the size of the station_explosion.dmi and screen1_full.dmi files. The rsc file will be about 8Meg smaller (so about the same size as it was a few months ago >_>)
If you want to reduce the size of a dmi, you can open it in GIMP, go to colours->posterize, then export it as png to import it back into byond.
If you need to split the image back into it's individual states you can set up guidelines and use filters->web->slice to slice it up into squares and automatically export it.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4996 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 1fbb83aad98705789aad73279c3c3d82c1571a5b