Removed the alertness level thing for carp/hostile mobs in general. Neat idea in theory, but in practice it just meant people could run up and beat carp to death before they'd even enter their hostile state. Bears still have it though.
Carp are slightly faster (in testing, I walked away from one as a borg, this seems a bit silly). This might not work well in practice though what with lag on a live server, so I'm guessing I might have to change this back.

Carp/syndicate also keep chasing their target as long as they're within 10 tiles, to compensate for them being slower than humans.

Removed the restriction on debraining changelings, this was a terrible "fix" to half-changelings and just lets you metagame.

Added a mineral var (metal) to floors to stop runtimes when you hit them with mining tools.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@5052 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: 7103cbfac01b8d755f6866fe1dfbe473f3c2ef70