-Changed turf creations, merging everything to only one proc -Fixes Issue 1085 - Artificers creating turfs with buggy lighting -Spells now will use the proper proc to create turfs
Due to the amount of files changed, this is just the 'part one'. The merging in turf.dm is not finished, mineral walls and buildmode are still using the old system. The creation of turfs on the game panel is not using the proper procs yet. Stay tuned for this changes, but now, sleep.


git-svn-id: http://tgstation13.googlecode.com/svn/trunk@5054 316c924e-a436-60f5-8080-3fe189b3f50e

Former-commit-id: dee7b74a45fc5dfa12d7e24647efa8ad744fb11a